+++ hydra-cluster/src/HydraNode.hs
@@ -20,6 +20,7 @@ import Data.Aeson.Types (Pair)
 import qualified Data.List as List
 import Data.Text (pack)
 import qualified Data.Text as T
+import Hydra.API.RestServer (DraftCommitTxRequest (..), DraftCommitTxResponse (..))
 import Hydra.Cluster.Faucet (FaucetLog)
 import Hydra.Cluster.Util (readConfigFile)
 import Hydra.ContestationPeriod (ContestationPeriod)
@@ -29,8 +30,8 @@ import Hydra.Logging (Tracer, Verbosity (..), traceWith)
 import Hydra.Network (Host (Host), NodeId (NodeId))
 import qualified Hydra.Network as Network
 import Hydra.Options (ChainConfig (..), LedgerConfig (..), RunOptions (..), defaultChainConfig, toArgs)
-import Network.HTTP.Conduit (HttpExceptionContent (ConnectionFailure), parseRequest)
-import Network.HTTP.Simple (HttpException (HttpExceptionRequest), Response, getResponseBody, getResponseStatusCode, httpBS)
+import Network.HTTP.Req (GET (..), HttpException, JsonResponse, NoReqBody (..), POST (..), ReqBodyJson (..), defaultHttpConfig, responseBody, runReq, (/:))
+import qualified Network.HTTP.Req as Req
 import Network.WebSockets (Connection, receiveData, runClient, sendClose, sendTextData)
 import System.FilePath ((<.>), (</>))
 import System.IO.Temp (withSystemTempDirectory)
@@ -111,7 +112,7 @@ waitForAllMatch delay nodes match = do
     failure "no clients to wait for"
   results <- forConcurrently nodes $ \n -> waitMatch delay n match
   case results of
-    [] -> failure $ "empty results, but " <> (show $ length nodes) <> " clients"
+    [] -> failure $ "empty results, but " <> show (length nodes) <> " clients"
     (r : rs) -> do
       unless (all (== r) rs) $
         failure $
@@ -161,23 +162,35 @@ waitForAll tracer delay nodes expected = do
         _ ->
           tryNext c msgs stillExpected
 
+-- | Create a commit tx using the hydra-node for later submission
+externalCommit :: HydraClient -> UTxO -> IO Tx
+externalCommit HydraClient{hydraNodeId} utxos =
+  runReq defaultHttpConfig request
+    <&> responseBody
+    >>= \DraftCommitTxResponse{commitTx} -> pure commitTx
+ where
+  request =
+    Req.req
+      POST
+      (Req.http "127.0.0.1" /: "commit")
+      (ReqBodyJson (DraftCommitTxRequest{utxos} :: DraftCommitTxRequest Tx))
+      (Proxy :: Proxy (JsonResponse (DraftCommitTxResponse Tx)))
+      (Req.port $ 4000 + hydraNodeId)
+
 getMetrics :: HasCallStack => HydraClient -> IO ByteString
 getMetrics HydraClient{hydraNodeId} = do
-  response <-
-    failAfter 3 $ queryNode hydraNodeId
-  when (getResponseStatusCode response /= 200) $ failure ("Request for Hydra-node metrics failed :" <> show (getResponseBody response))
-  pure $ getResponseBody response
-
-queryNode :: Int -> IO (Response ByteString)
-queryNode nodeId =
-  parseRequest ("http://127.0.0.1:" <> show (6000 + nodeId) <> "/metrics") >>= loop
+  failAfter 3 $
+    try (runReq defaultHttpConfig request) >>= \case
+      Left (e :: HttpException) -> failure $ "Request for hydra-node metrics failed: " <> show e
+      Right body -> pure $ Req.responseBody body
  where
-  loop req =
-    httpBS req `catch` onConnectionFailure (loop req)
-
-  onConnectionFailure cont = \case
-    (HttpExceptionRequest _ (ConnectionFailure _)) -> threadDelay 100_000 >> cont
-    e -> throwIO e
+  request =
+    Req.req
+      GET
+      (Req.http "127.0.0.1" /: "metrics")
+      NoReqBody
+      Req.bsResponse
+      (Req.port $ 6000 + hydraNodeId)
 
 data EndToEndLog
   = NodeStarted {nodeId :: Int}