+++ hydra-cluster/test/Test/EndToEndSpec.hs
@@ -9,7 +9,7 @@ import Hydra.Prelude
 import Test.Hydra.Prelude
 
 import qualified Cardano.Api.UTxO as UTxO
-import CardanoClient (QueryPoint (..), queryGenesisParameters, queryTip, queryTipSlotNo, waitForUTxO)
+import CardanoClient (QueryPoint (..), queryGenesisParameters, queryTip, queryTipSlotNo, submitTx, waitForUTxO)
 import CardanoNode (RunningNode (..), generateCardanoKey, withCardanoNodeDevnet)
 import Control.Concurrent.STM (newTVarIO, readTVarIO)
 import Control.Concurrent.STM.TVar (modifyTVar')
@@ -38,6 +38,7 @@ import Hydra.Cardano.Api (
   lovelaceToValue,
   mkVkAddress,
   serialiseAddress,
+  signTx,
   writeFileTextEnvelope,
   pattern TxValidityLowerBound,
  )
@@ -65,7 +66,7 @@ import Hydra.Cluster.Scenarios (
   headIsInitializingWith,
   restartedNodeCanAbort,
   restartedNodeCanObserveCommitTx,
-  singlePartyCommitsFromExternal,
+  singlePartyCommitsUsingFuel,
   singlePartyHeadFullLifeCycle,
  )
 import Hydra.Cluster.Util (chainConfigFor, keysFor)
@@ -80,6 +81,7 @@ import Hydra.Party (deriveParty)
 import HydraNode (
   EndToEndLog (..),
   HydraClient,
+  externalCommit,
   getMetrics,
   input,
   output,
@@ -132,11 +134,11 @@ spec = around showLogsOnFailure $ do
           withCardanoNodeDevnet (contramap FromCardanoNode tracer) tmpDir $ \node ->
             publishHydraScriptsAs node Faucet
               >>= timedTx tmpDir tracer node
-      it "commits from external wallet" $ \tracer -> do
-        withClusterTempDir "single-commits-from-external" $ \tmpDir -> do
+      it "commits using fuel" $ \tracer -> do
+        withClusterTempDir "commits-using-fuel" $ \tmpDir -> do
           withCardanoNodeDevnet (contramap FromCardanoNode tracer) tmpDir $ \node ->
             publishHydraScriptsAs node Faucet
-              >>= singlePartyCommitsFromExternal tracer tmpDir node
+              >>= singlePartyCommitsUsingFuel tracer tmpDir node
 
     describe "three hydra nodes scenario" $ do
       it "inits a Head, processes a single Cardano transaction and closes it again" $ \tracer ->
@@ -146,7 +148,7 @@ spec = around showLogsOnFailure $ do
               hydraScriptsTxId <- publishHydraScriptsAs node Faucet
               initAndClose tmpDir tracer 1 hydraScriptsTxId node
 
-      it "inits a Head and closes it immediately " $ \tracer ->
+      it "inits a Head and closes it immediately" $ \tracer ->
         failAfter 60 $
           withClusterTempDir "three-init-close-immediately" $ \tmpDir -> do
             let clusterIx = 0
@@ -177,11 +179,15 @@ spec = around showLogsOnFailure $ do
                     headIsInitializingWith (Set.fromList [alice, bob, carol])
 
                 -- Get some UTXOs to commit to a head
-                committedUTxOByAlice <- seedFromFaucet node aliceCardanoVk aliceCommittedToHead Normal (contramap FromFaucet tracer)
-                committedUTxOByBob <- seedFromFaucet node bobCardanoVk bobCommittedToHead Normal (contramap FromFaucet tracer)
-                send n1 $ input "Commit" ["utxo" .= committedUTxOByAlice]
-                send n2 $ input "Commit" ["utxo" .= committedUTxOByBob]
-                send n3 $ input "Commit" ["utxo" .= Object mempty]
+                (aliceExternalVk, aliceExternalSk) <- generate genKeyPair
+                committedUTxOByAlice <- seedFromFaucet node aliceExternalVk aliceCommittedToHead Normal (contramap FromFaucet tracer)
+                externalCommit n1 committedUTxOByAlice <&> signTx aliceExternalSk >>= submitTx node
+
+                (bobExternalVk, bobExternalSk) <- generate genKeyPair
+                committedUTxOByBob <- seedFromFaucet node bobExternalVk bobCommittedToHead Normal (contramap FromFaucet tracer)
+                externalCommit n2 committedUTxOByBob <&> signTx bobExternalSk >>= submitTx node
+
+                externalCommit n3 mempty >>= submitTx node
 
                 let u0 = committedUTxOByAlice <> committedUTxOByBob
 
@@ -252,7 +258,7 @@ spec = around showLogsOnFailure $ do
           withCardanoNodeDevnet (contramap FromCardanoNode tracer) tmp $ \node@RunningNode{nodeSocket, networkId} -> do
             hydraScriptsTxId <- publishHydraScriptsAs node Faucet
 
-            (aliceCardanoVk, aliceCardanoSk) <- keysFor Alice
+            (aliceCardanoVk, _aliceCardanoSk) <- keysFor Alice
             (bobCardanoVk, _bobCardanoSk) <- keysFor Bob
 
             seedFromFaucet_ node aliceCardanoVk 100_000_000 Fuel (contramap FromFaucet tracer)
@@ -279,9 +285,13 @@ spec = around showLogsOnFailure $ do
                 send n1 $ input "Init" []
                 headId <- waitForAllMatch 10 [n1, n2] $ headIsInitializingWith (Set.fromList [alice, bob])
 
-                committedUTxOByAlice <- seedFromFaucet node aliceCardanoVk aliceCommittedToHead Normal (contramap FromFaucet tracer)
-                send n1 $ input "Commit" ["utxo" .= committedUTxOByAlice]
-                send n2 $ input "Commit" ["utxo" .= Object mempty]
+                (aliceExternalVk, aliceExternalSk) <- generate genKeyPair
+                committedUTxOByAlice <- seedFromFaucet node aliceExternalVk aliceCommittedToHead Normal (contramap FromFaucet tracer)
+                externalCommit n1 committedUTxOByAlice <&> signTx aliceExternalSk >>= submitTx node
+
+                (bobExternalVk, _bobExternalSk) <- generate genKeyPair
+                externalCommit n2 mempty >>= submitTx node
+
                 waitFor tracer 10 [n1, n2] $ output "HeadIsOpen" ["utxo" .= committedUTxOByAlice, "headId" .= headId]
 
                 -- Create an arbitrary transaction using some input.
@@ -289,8 +299,8 @@ spec = around showLogsOnFailure $ do
                 let Right tx =
                       mkSimpleTx
                         firstCommittedUTxO
-                        (inHeadAddress bobCardanoVk, lovelaceToValue paymentFromAliceToBob)
-                        aliceCardanoSk
+                        (inHeadAddress bobExternalVk, lovelaceToValue paymentFromAliceToBob)
+                        aliceExternalSk
                 send n1 $ input "NewTx" ["transaction" .= tx]
 
                 waitMatch 10 n1 $ \v -> do
@@ -361,7 +371,7 @@ spec = around showLogsOnFailure $ do
                     output "HeadIsAborted" ["utxo" .= Object mempty, "headId" .= headIdAliceAndBob]
 
                   -- Alice should be able to continue working with her Head
-                  send n1 $ input "Commit" ["utxo" .= Object mempty]
+                  externalCommit n1 mempty >>= submitTx node
                   waitFor tracer 10 [n1] $
                     output "HeadIsOpen" ["utxo" .= Object mempty, "headId" .= headIdAliceOnly]
 
@@ -504,7 +514,7 @@ waitForLog delay nodeOutput failureMessage predicate = do
 
 timedTx :: FilePath -> Tracer IO EndToEndLog -> RunningNode -> TxId -> IO ()
 timedTx tmpDir tracer node@RunningNode{networkId, nodeSocket} hydraScriptsTxId = do
-  (aliceCardanoVk, aliceCardanoSk) <- keysFor Alice
+  (aliceCardanoVk, _) <- keysFor Alice
   let aliceSk = generateSigningKey "alice-timed"
   let alice = deriveParty aliceSk
   let contestationPeriod = UnsafeContestationPeriod 2
@@ -521,8 +531,10 @@ timedTx tmpDir tracer node@RunningNode{networkId, nodeSocket} hydraScriptsTxId =
         headIsInitializingWith (Set.fromList [alice])
 
     -- Get some UTXOs to commit to a head
-    committedUTxOByAlice <- seedFromFaucet node aliceCardanoVk aliceCommittedToHead Normal (contramap FromFaucet tracer)
-    send n1 $ input "Commit" ["utxo" .= committedUTxOByAlice]
+    (aliceExternalVk, aliceExternalSk) <- generate genKeyPair
+    committedUTxOByAlice <- seedFromFaucet node aliceExternalVk aliceCommittedToHead Normal (contramap FromFaucet tracer)
+    externalCommit n1 committedUTxOByAlice <&> signTx aliceExternalSk >>= submitTx node
+
     waitFor tracer 3 [n1] $ output "HeadIsOpen" ["utxo" .= committedUTxOByAlice, "headId" .= headId]
 
     -- Acquire a current point in time
@@ -543,8 +555,8 @@ timedTx tmpDir tracer node@RunningNode{networkId, nodeSocket} hydraScriptsTxId =
         Right tx =
           mkRangedTx
             firstCommittedUTxO
-            (inHeadAddress aliceCardanoVk, lovelaceToValue lovelaceToSend)
-            aliceCardanoSk
+            (inHeadAddress aliceExternalVk, lovelaceToValue lovelaceToSend)
+            aliceExternalSk
             (Just $ TxValidityLowerBound futureSlot, Nothing)
 
     -- First submission: invalid
@@ -567,7 +579,7 @@ timedTx tmpDir tracer node@RunningNode{networkId, nodeSocket} hydraScriptsTxId =
 
 initAndClose :: FilePath -> Tracer IO EndToEndLog -> Int -> TxId -> RunningNode -> IO ()
 initAndClose tmpDir tracer clusterIx hydraScriptsTxId node@RunningNode{nodeSocket, networkId} = do
-  aliceKeys@(aliceCardanoVk, aliceCardanoSk) <- generate genKeyPair
+  aliceKeys@(aliceCardanoVk, _) <- generate genKeyPair
   bobKeys@(bobCardanoVk, _) <- generate genKeyPair
   carolKeys@(carolCardanoVk, _) <- generate genKeyPair
 
@@ -599,11 +611,16 @@ initAndClose tmpDir tracer clusterIx hydraScriptsTxId node@RunningNode{nodeSocke
         headIsInitializingWith (Set.fromList [alice, bob, carol])
 
     -- Get some UTXOs to commit to a head
-    committedUTxOByAlice <- seedFromFaucet node aliceCardanoVk aliceCommittedToHead Normal (contramap FromFaucet tracer)
-    committedUTxOByBob <- seedFromFaucet node bobCardanoVk bobCommittedToHead Normal (contramap FromFaucet tracer)
-    send n1 $ input "Commit" ["utxo" .= committedUTxOByAlice]
-    send n2 $ input "Commit" ["utxo" .= committedUTxOByBob]
-    send n3 $ input "Commit" ["utxo" .= Object mempty]
+    (aliceExternalVk, aliceExternalSk) <- generate genKeyPair
+    committedUTxOByAlice <- seedFromFaucet node aliceExternalVk aliceCommittedToHead Normal (contramap FromFaucet tracer)
+    externalCommit n1 committedUTxOByAlice <&> signTx aliceExternalSk >>= submitTx node
+
+    (bobExternalVk, bobExternalSk) <- generate genKeyPair
+    committedUTxOByBob <- seedFromFaucet node bobExternalVk bobCommittedToHead Normal (contramap FromFaucet tracer)
+    externalCommit n2 committedUTxOByBob <&> signTx bobExternalSk >>= submitTx node
+
+    externalCommit n3 mempty >>= submitTx node
+
     waitFor tracer 10 [n1, n2, n3] $ output "HeadIsOpen" ["utxo" .= (committedUTxOByAlice <> committedUTxOByBob), "headId" .= headId]
 
     -- NOTE(AB): this is partial and will fail if we are not able to generate a payment
@@ -611,8 +628,8 @@ initAndClose tmpDir tracer clusterIx hydraScriptsTxId node@RunningNode{nodeSocke
     let Right tx =
           mkSimpleTx
             firstCommittedUTxO
-            (inHeadAddress bobCardanoVk, lovelaceToValue paymentFromAliceToBob)
-            aliceCardanoSk
+            (inHeadAddress bobExternalVk, lovelaceToValue paymentFromAliceToBob)
+            aliceExternalSk
     send n1 $ input "NewTx" ["transaction" .= tx]
     waitFor tracer 10 [n1, n2, n3] $
       output "TxValid" ["transaction" .= tx, "headId" .= headId]
@@ -625,7 +642,7 @@ initAndClose tmpDir tracer clusterIx hydraScriptsTxId node@RunningNode{nodeSocke
             [
               ( TxIn (txId tx) (toEnum 0)
               , object
-                  [ "address" .= String (serialiseAddress $ inHeadAddress bobCardanoVk)
+                  [ "address" .= String (serialiseAddress $ inHeadAddress bobExternalVk)
                   , "value" .= object ["lovelace" .= int paymentFromAliceToBob]
                   , "datum" .= Null
                   , "datumhash" .= Null
@@ -636,7 +653,7 @@ initAndClose tmpDir tracer clusterIx hydraScriptsTxId node@RunningNode{nodeSocke
             ,
               ( TxIn (txId tx) (toEnum 1)
               , object
-                  [ "address" .= String (serialiseAddress $ inHeadAddress aliceCardanoVk)
+                  [ "address" .= String (serialiseAddress $ inHeadAddress aliceExternalVk)
                   , "value" .= object ["lovelace" .= int (aliceCommittedToHead - paymentFromAliceToBob)]
                   , "datum" .= Null
                   , "datumhash" .= Null